+++ app/uk/gov/hmrc/nationalinsurancecontributionandcreditsapi/models/Request.scala
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.bereavementsupportpaymentapi.models
+package uk.gov.hmrc.nationalinsurancecontributionandcreditsapi.models
 import com.google.inject.Inject
-import uk.gov.hmrc.bereavementsupportpaymentapi.utils.{AdditionalHeaderNames, RequestParams, Validator}
+import uk.gov.hmrc.nationalinsurancecontributionandcreditsapi.utils.{AdditionalHeaderNames, RequestParams, Validator}
 import java.time.LocalDate
 import java.time.format.{DateTimeFormatter, DateTimeParseException}
