+++ app/uk/gov/hmrc/nationalinsurancecontributionandcreditsapi/models/package.scala
@@ -14,18 +14,10 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.nationalinsurancecontributionandcreditsapi.controllers
+package uk.gov.hmrc.nationalinsurancecontributionandcreditsapi
 
-import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
-import play.api.mvc.{Action, AnyContent, ControllerComponents}
-import javax.inject.{Inject, Singleton}
-import scala.concurrent.Future
+import uk.gov.hmrc.nationalinsurancecontributionandcreditsapi.models.errors.Errors
 
-@Singleton()
-class MicroserviceHelloWorldController @Inject()(cc: ControllerComponents)
-    extends BackendController(cc) {
-
-  def hello(): Action[AnyContent] = Action.async { implicit request =>
-    Future.successful(Ok("Hello world"))
-  }
+package object models {
+  type HIPOutcome = Either[Errors, Response]
 }
