+++ app/uk/gov/hmrc/nationalinsurancecontributionandcreditsapi/utils/RequestParams.scala
@@ -14,17 +14,12 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.bereavementsupportpaymentapi.config
+package uk.gov.hmrc.nationalinsurancecontributionandcreditsapi.utils
 
-import javax.inject.{Inject, Singleton}
-import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
-
-@Singleton
-class AppConfig @Inject()(config: ServicesConfig) {
-
-  private val hipServicePrefix = "microservice.services.hip"
-
-  lazy val hipBaseUrl: String = config.baseUrl("hip")
-  lazy val hipToken: String = config.getString(s"$hipServicePrefix.token")
-  lazy val hipEnvironment: String = config.getString(s"$hipServicePrefix.env")
+object RequestParams {
+  val NINO = "nino"
+  val FORENAME = "forename"
+  val SURNAME = "surname"
+  val DATE_OF_BIRTH = "dateOfBirth"
+  val DATE_RANGE = "dateRange"
 }
