--- test/uk/gov/hmrc/nationalinsurancecontributionandcreditsapi/controllers/MicroserviceHelloWorldControllerSpec.scala
@@ -1,20 +1,0 @@
-package uk.gov.hmrc.nationalinsurancecontributionandcreditsapi.controllers
-
-import org.scalatest.matchers.should.Matchers
-import org.scalatest.wordspec.AnyWordSpec
-import play.api.http.Status
-import play.api.test.Helpers._
-import play.api.test.{FakeRequest, Helpers}
-
-class MicroserviceHelloWorldControllerSpec extends AnyWordSpec with Matchers {
-
-  private val fakeRequest = FakeRequest("GET", "/")
-  private val controller = new MicroserviceHelloWorldController(Helpers.stubControllerComponents())
-
-  "GET /" should {
-    "return 200" in {
-      val result = controller.hello()(fakeRequest)
-      status(result) shouldBe Status.OK
-    }
-  }
-}
